Youtube support
The big issue with youtube support is that position is only updated if
mouse is over youtube video and I have not found a way to reverse
engineer it to always work.
I think now I am going to work on getting the plugin side of things to
be a little nicer before moving on to spotify.