Fixes for youtube and soundcloud
Youtube now no longer does the mouse move trick when mousing down, I
will also make it not when fullscreen in a future update if I dont
remove it by then.
Fixed a bug where if you moused over an artist it would send that image
to the plugin until you moused off.
Also compliant with new changes to the WebNowPlaying plugin.