Fixed an issue where youtube videos in new tab
Youtube videos when opened in a new tab do not start playing till you
give them focus. However the video element is marked as playing. By also
checking if any of the video has been played as well it will no longer
hijack WebNowPlaying to show it instead of the one you are actually
watching.