Added SetPosition for Amazon Music
Amazon music's web interface seems poorly maintained and has some issues
where it seems to think it has not buffered a specific song of mine.
There is no way around this as it does not work under normal operation
let alone more complex.
Also the volume slider seems to have some issues as well so after trying
to figure out why it throws away clicks and work around it for an hour I
have decided it is not worth it. I left the code in but commented it out
and had it throw an error.