Fixed Deezer support...mostly
Dear god their new layout is dogshit under the hood, it uses tables in
some places. The system I had to write for their rating system is a
clusterfuck

Sadly volume value is now just always going to be 100 (The mouse event
for it leaves it open half the time, even when just using it normally),
and both the volume and progress no longer can be controlled. Blame
their shit site