Various fixes
Git decided and somehow all my commits for this repo ended up thinking was for a private repo I have thus causing git history to get completely messed for this repo so sadly I had to blow everything away and just copy the files over.

In a nutshell:
Pocketcasts: Fixed artwork.

Spotify: They renamed some elements internally so just updated to match it, it also now gets the higher res album art if it is there (Need to cache it though) and album is still broken.

Youtube Music: Fixed album art so it downloads again also fixed it so that the album is correct and handles youtube videos better