Fix to show only "HD" Album art.
I monitored the network log when loading low and high res art work, and they seem the same, except the first few characters. I tried replacing the art's `src` attribute to match the "HD" version every time, and it seems to work for me. **PLEASE TEST THIS WITH YOUR MACHINE, I'M NOT SURE IF IT WILL DEFINITELY WORK OR NOT...** Also not sure how it might effect the rate limit.