Better WS error suppression
I no check if the websocket is open before sending any info over to the server.

This was not a big deal as the only reason it would pass the other checks is during the page unload/reload. Due to wanting to ensure that it does not happen as much as possible the check is done before every send (I almost made it its own function but felt that was overkill in this case)